RetryPatcher
