+++ src/greptimeai/patcher/__init__.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class Patcher(ABC):
+    @abstractmethod
+    def patch(self):
+        pass
