+++ src/greptimeai/patcher/openai/retry.py
@@ -1,0 +1,44 @@
+import functools
+from typing import Union
+
+from openai import AsyncOpenAI, OpenAI
+from typing_extensions import override
+
+from greptimeai import logger
+from greptimeai.patchee import Patchee
+from greptimeai.patchee.openai.retry import RetryPatchees
+from greptimeai.tracker import Tracker
+
+from .base import _OpenaiPatcher
+
+
+class _RetryPatcher(_OpenaiPatcher):
+    def __init__(
+        self,
+        tracker: Tracker,
+        client: Union[OpenAI, AsyncOpenAI, None] = None,
+    ):
+        patchees = RetryPatchees(client=client)
+        super().__init__(tracker=tracker, patchees=patchees, client=client)
+
+    @override
+    def patch_one(self, patchee: Patchee):
+        func = patchee.get_unwrapped_func()
+        if not func:
+            return
+
+        if self.is_async:
+            pass
+        else:
+
+            @functools.wraps(func)
+            def wrapper(*args, **kwargs):
+                try:
+                    logger.info(f"in retry_patcher { args = }")
+                    logger.info(f"in retry_patcher { kwargs = }")
+                    resp = func(*args, **kwargs)
+                except Exception as e:
+                    raise e
+                return resp
+
+            patchee.wrap_func(wrapper)
